airflow/requirements.txt
astronomer-cosmos==1.10.0
dbt-core
dbt-snowflake
apache-airflow-providers-snowflake
snowflake-connector-python
dbt-postgres
# dbt-duckdb